requirement.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
beautifulsoup4==4.6.0
bs4==0.0.1
bson==0.5.5
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
Django==1.11.13
django-crispy-forms==1.7.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jieba==0.39
kiwisolver==1.0.1
lxml==4.2.1
matplotlib==2.2.2
numpy==1.14.3
pandas==0.22.0
parsel==1.4.0
Pillow==5.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pymongo==3.6.1
PyMySQL==0.8.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
queuelib==1.5.0
requests==2.18.4
scikit-learn==0.19.1
scipy==1.1.0
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
sklearn==0.0
statsmodels==0.9.0
texttable==1.2.1
Twisted==18.4.0
urllib3==1.22
w3lib==1.19.0
xlwt==1.3.0
zope.interface==4.5.0
